Move mock MM to its own file and accept address & chainId.
